% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_selection.R
\name{backward_selection_saemix}
\alias{backward_selection_saemix}
\title{Boundary LRT backward selection for random effects (saemix)}
\usage{
backward_selection_saemix(
  par_names,
  saemix_data,
  saemix_model_fun,
  psi0,
  alpha = 0.01,
  seed = 990202,
  saemix_control = list(),
  initial_fit = NULL
)
}
\arguments{
\item{par_names}{character vector of parameter names (log-scale)}

\item{saemix_data}{saemixData}

\item{saemix_model_fun}{function(psi,id,xidep) returning y}

\item{psi0}{1xP matrix of initial fixed effects (log-scale)}

\item{alpha}{boundary LRT alpha, default 0.01}

\item{seed}{Integer random seed.}

\item{saemix_control}{Named SAEMIX control options.}

\item{initial_fit}{Optional already-fitted full random-effect model.}
}
\value{
list(selected, covariance.model)
}
\description{
Boundary LRT backward selection for random effects (saemix)
}
\keyword{internal}
